Fixed item/slot enumeration and string formatting.
Differentiated "slots" (a place for an item on a character) and "items" (the
valid items the enchant can appear on). You can still filter by slot, but the
complete list of valid item types (raw, for the moment) appears in the details.

Fixed {ability} and {deflect} to refer to a more specific but still class-
independent string.

Fixed HTML formatting in all enchant descriptions.